requirements.txt
alembic==1.4.3
asgiref==3.3.1
beautifulsoup4==4.9.3
boto3==1.17.53
botocore==1.20.53
certifi==2018.11.29
chardet==3.0.4
dataset==1.3.2
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.6
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-crispy-forms==1.9.2
django-import-export==2.4.0
django-storages==1.11.1
django-translation-flags==1.0.6
django-translations==1.2.0
docopt==0.6.2
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
jdcal==1.4.1
jmespath==0.10.0
Mako==1.1.3
MarkupPy==1.14
MarkupSafe==1.1.1
numpy==1.20.2
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.2.4
pandas-stubs==1.1.0.2
Pillow==8.2.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-gettext==4.0
pytz==2021.1
PyYAML==5.3.1
requests==2.24.0
res==0.1.7
s3transfer==0.3.7
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
tablib==2.0.0
urllib3==1.25.11
virtualenv==16.4.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
